+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 ###TODO: 
 
 1. ActorModel的status字段需要写一个迁移程序
-2. 修复redis key "wating_clear_processors" 单词错误迁移程序
+2. 修复redis key "waiting_clear_processors" 单词错误迁移程序